[ADDED] Per-User rate limiting
This is a very basic and not necessarily accurate per-user rate
limiting. Both ingress and egress are counted toward an user limit.
Since delivery is done synchronously from the producer readLoop,
producer may be blocked by a consumer rate limit even if that producer
has no rate limit defined or its rate limit is higher than the
consumer's.
The changes when running without rate limit should not have any
performance impact. When rate limiting is enabled, but limit is
very high, a performance degradation is possible (cost of 3 atomic
operations).